rest cleaned. body construction moved into api
