+++ barf/index.js
@@ -1,7 +1,9 @@
+const BigNumber = require('bignumber.js');
 const rest = require('./rest_7.js')
 
 // TODO add functionality to assert HERE
 
 module.exports = {
+  BigNumber,
   rest,
 }
