+++ barf/test/index.test.js
@@ -214,25 +214,3 @@ describe('user', function() {
     assert.equal(address, user.address, 'user is valid eth address')
   })
 })
-
-describe('include rest', function() {
-  this.timeout(config.timeout)
-  it('testAsync', async () => {
-    const args = { a: 'b' }
-    const result = await rest.testAsync(args)
-    assert.deepEqual(result, args, 'test async')
-  })
-
-  it('testPromise', async () => {
-    const args = { success: true }
-    const result = await rest.testPromise(args)
-    assert.deepEqual(result, args, 'test promise')
-
-    args.success = false
-    try {
-      await rest.testPromise(args)
-    } catch (err) {
-      assert.deepEqual(err, args, 'test promise')
-    }
-  })
-})
